+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/datasource/EnsemblJannovarDataFactory.java
@@ -4,7 +4,6 @@ import org.ini4j.Profile.Section;
 
 import com.google.common.collect.ImmutableList;
 
-import de.charite.compbio.jannovar.JannovarOptions;
 import de.charite.compbio.jannovar.impl.parse.EnsemblParser;
 import de.charite.compbio.jannovar.impl.parse.TranscriptParseException;
 import de.charite.compbio.jannovar.io.ReferenceDictionary;
@@ -27,7 +26,7 @@ final class EnsemblJannovarDataFactory extends JannovarDataFactory {
 	 * @param iniSection
 	 *            {@link Section} with configuration from INI file
 	 */
-	public EnsemblJannovarDataFactory(JannovarOptions options, EnsemblDataSource dataSource, Section iniSection,
+	public EnsemblJannovarDataFactory(DatasourceOptions options, EnsemblDataSource dataSource, Section iniSection,
 			boolean printProgressBars) {
 		super(options, dataSource, iniSection);
 	}
@@ -35,7 +34,7 @@ final class EnsemblJannovarDataFactory extends JannovarDataFactory {
 	@Override
 	protected ImmutableList<TranscriptModel> parseTranscripts(ReferenceDictionary refDict, String targetDir)
 			throws TranscriptParseException {
-		return new EnsemblParser(refDict, targetDir, iniSection, options.printProgressBars).run();
+		return new EnsemblParser(refDict, targetDir, iniSection, options.doPrintProgressBars()).run();
 	}
 
 }